added xml generation capability
it can now generate an executable jar. the jar can be run providing an
input and an output directory. the input directory needs to contain all
the metadata files with 9 lines of metadata with extension '.metadata'.
it will then create iso19115 xmls in the output directory with those
metadata and  given default data in defaults.properties